% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irc_regression.R
\name{irc_regression}
\alias{irc_regression}
\title{Irrepresentable Condition: Regression}
\usage{
irc_regression(X, which_nonzero)
}
\arguments{
\item{X}{A matrix of dimensions \emph{n} (observations) by \emph{p} (variables)}

\item{which_nonzero}{Numeric vector with the location of the nonzero relations
(a.k.a., the active set)}
}
\value{
1 - infinity norm (negative the IRC is violated)
}
\description{
Irrepresentable Condition: Regression
}
\examples{
\donttest{
# data
# note: irc_met (block diagonal; 1st 10 active)
cors <- IRCcheck:::irc_met
X <- MASS::mvrnorm(2500, rep(0, 20), Sigma = cors, empirical = TRUE)
 
# check IRC
irc_regression(X, which_nonzero = 1:10)  

# generate data
y <- X \%*\% c(rep(1,10), rep(0, 10)) + rnorm(2500)

fit <- glmnet::glmnet(X, y, lambda = seq(10, 0.01, length.out = 400))

# plot
plot(fit, xvar = "lambda")

}
}
