% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symm_mat.R
\name{symm_mat}
\alias{symm_mat}
\title{Symmetric Matrix

Copy the upper triangular of a matrix into the lower triangular portion of the matrix.}
\usage{
symm_mat(x)
}
\arguments{
\item{x}{A matrix of dimensions \emph{p} by \emph{p}.}
}
\value{
A matrix
}
\description{
Symmetric Matrix

Copy the upper triangular of a matrix into the lower triangular portion of the matrix.
}
\examples{
adj <- matrix(sample(0:1, size = 25, replace = TRUE), 5, 5)
symm_mat(adj)

}
